requirements.txt
# Runpod dependencies
runpod~=1.7.0

# ComfyUI dependencies
torch
torchsde
torchvision
torchaudio
numpy>=1.25.0
einops
transformers>=4.28.1
tokenizers>=0.13.3
sentencepiece
safetensors>=0.4.2
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil

# ComfyUI testing dependencies:
websocket-client

# ComfyUI Non essential dependencies:
kornia>=0.7.1
spandrel
soundfile